Cambios en la funcionalidad de frmFactura, se agregan los combobox y se corriguen errores

+++ LPCFacturas/ProyectoBugs/DataAccessLayer/ClienteDao.cs
@@ -78,7 +78,7 @@ namespace LPCFacturas.DataAccessLayer
             Cliente oCliente = new Cliente
             {
                 Id_cliente = Convert.ToInt32(row["id_cliente"].ToString()),
-                Cuit = Convert.ToInt32(row["cuit"].ToString()),
+                Cuit = Convert.ToInt64(row["cuit"].ToString()),
                 Razon_social = row["razon_social"].ToString(),
                 Calle = row["calle"].ToString(),
                 Numero = Convert.ToInt32(row["numero"].ToString()),
